src/requirements.txt
ansi-escapes==0.1.1
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudscraper==1.2.71
codeforces==0.1.1
colorama==0.4.6
comm==0.1.3
comtypes==1.2.0
cryptography==41.0.1
dateparser==1.1.8
debugpy==1.6.7
decorator==5.1.1
dxcam==0.0.5
easyocr==1.7.0
Eel==0.16.0
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.12.2
Flask==2.3.2
future==0.18.3
getkey==0.6.5
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
imageio==2.31.1
imutils==0.5.4
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
numpy==1.25.1
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
PyAudio==0.2.13
pyclipper==1.3.0.post4
pycparser==2.21
pyee==9.1.0
Pygments==2.15.1
pygrabber==0.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-ffmpeg==2.0.4
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
scikit-image==0.21.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
socialblade==1.0.5.1
soupsieve==2.4.1
stack-data==0.6.2
sty==1.0.4
sympy==1.12
tifffile==2023.7.10
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.1
tzdata==2023.3
tzlocal==5.0.1
Unidecode==1.3.6
urllib3==2.0.3
wcwidth==0.2.6
websocket-client==1.6.1
Werkzeug==2.3.6
whichcraft==0.6.1
youtube-dl==2021.12.17
zope.event==5.0
zope.interface==6.0